Move files from Wiktionary utils to WD utils - delete Wiktionary dir
