--- docs/source/scribe_data/wiktionary/index.rst
@@ -1,6 +1,0 @@
-wiktionary/
-===========
-
-`View code on Github <https://github.com/scribe-org/Scribe-Data/tree/main/src/scribe_data/wiktionary>`_
-
-Scribe-Data will eventually be using data from Wiktionaries for features such as deriving translations of words as well as synonyms and antonyms.
